Port code actions tests to the new framework
Closes #20371
